backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI RESTful backend with Langchain, LangGraph, and Ollama integration for therapist chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.22.1",
    "httpx>=0.25.0",
    "langchain>=1.2.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langchain-ollama>=1.0.1",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
